Reorganize BaseService-related modules
Ref #2698
