+++ core/base-service/base-non-memory-caching.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const makeBadge = require('../gh-badges/lib/make-badge')
-const { makeSend } = require('../lib/result-sender')
+const makeBadge = require('../../gh-badges/lib/make-badge')
+const { makeSend } = require('../../lib/result-sender')
 const BaseService = require('./base')
 const { setCacheHeaders } = require('./cache-headers')
