+++ core/base-service/base-static.js
@@ -1,8 +1,8 @@
 'use strict'
 
-const makeBadge = require('../gh-badges/lib/make-badge')
-const { makeSend } = require('../lib/result-sender')
-const analytics = require('../core/server/analytics')
+const makeBadge = require('../../gh-badges/lib/make-badge')
+const { makeSend } = require('../../lib/result-sender')
+const analytics = require('../server/analytics')
 const BaseService = require('./base')
 const {
   serverHasBeenUpSinceResourceCached,
