+++ core/base-service/base-svg-scraping.js
