+++ core/base-service/base-xml.js
