+++ core/base-service/base-xml.spec.js
