+++ core/base-service/errors.js
