+++ lib/deprecation-helpers.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { Deprecated } = require('../services/errors')
+const { Deprecated } = require('../core/base-service/errors')
 
 function enforceDeprecation(effectiveDate) {
   if (Date.now() >= effectiveDate.getTime()) {
