+++ lib/deprecation-helpers.spec.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { expect } = require('chai')
-const { Deprecated } = require('../services/errors')
+const { Deprecated } = require('../core/base-service/errors')
 const { enforceDeprecation } = require('./deprecation-helpers')
 
 describe('enforceDeprecation', function() {
