+++ lib/error-helper.js
@@ -4,7 +4,7 @@ const {
   NotFound,
   InvalidResponse,
   Inaccessible,
-} = require('../services/errors')
+} = require('../core/base-service/errors')
 const defaultErrorMessages = {
   404: 'not found',
 }
