+++ lib/error-helper.spec.js
@@ -1,12 +1,12 @@
 'use strict'
 
 const { expect } = require('chai')
-const { checkErrorResponse } = require('./error-helper')
 const {
   NotFound,
   InvalidResponse,
   Inaccessible,
-} = require('../services/errors')
+} = require('../core/base-service/errors')
+const { checkErrorResponse } = require('./error-helper')
 
 describe('Standard Error Handler', function() {
   it('makes inaccessible badge', function() {
