+++ lib/regular-update.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { Inaccessible, InvalidResponse } = require('../services/errors')
+const { Inaccessible, InvalidResponse } = require('../core/base-service/errors')
 
 // Map from URL to { timestamp: last fetch time, data: data }.
 let regularUpdateCache = Object.create(null)
