+++ lib/request-handler.js
@@ -12,8 +12,8 @@ const {
   Inaccessible,
   InvalidResponse,
   ShieldsRuntimeError,
-} = require('../services/errors')
-const { setCacheHeaders } = require('../services/cache-headers')
+} = require('../core/base-service/errors')
+const { setCacheHeaders } = require('../core/base-service/cache-headers')
 const { makeBadgeData: getBadgeData } = require('./badge-data')
 const { makeSend } = require('./result-sender')
