+++ services/azure-devops/azure-devops-release.service.js
@@ -1,8 +1,8 @@
 'use strict'
 
-const BaseSvgService = require('../base-svg-scraping')
-const { keywords, fetch } = require('./azure-devops-helpers')
 const { renderBuildStatusBadge } = require('../../lib/build-status')
+const { BaseSvgScrapingService } = require('..')
+const { keywords, fetch } = require('./azure-devops-helpers')
 
 const documentation = `
 <p>
@@ -22,7 +22,7 @@ const documentation = `
 </p>
 `
 
-module.exports = class AzureDevOpsRelease extends BaseSvgService {
+module.exports = class AzureDevOpsRelease extends BaseSvgScrapingService {
   static get category() {
     return 'build'
   }
