+++ services/dynamic-common.js
@@ -3,7 +3,7 @@
 const Joi = require('joi')
 const validate = require('../core/base-service/validate')
 const { toArray } = require('../lib/badge-data')
-const { InvalidResponse } = require('./errors')
+const { InvalidResponse } = require('.')
 
 const individualValueSchema = Joi.alternatives()
   .try(Joi.string(), Joi.number())
