+++ test-fixtures/valid-class.fixture.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const BaseJsonService = require('../services/base-json')
+const BaseJsonService = require('../core/base-service/base-json')
 
 class GoodService extends BaseJsonService {
   static get category() {
